+++ app/src/main/java/org/simple/clinic/home/overdue/OverdueEvent.kt
@@ -13,7 +13,7 @@ data class CallPatientClicked(val patientUuid: UUID) : OverdueEvent() {
   override val analyticsName = "Overdue Screen:Call Patient clicked"
 }
 
-data class PatientNameClicked(val patientUuid: UUID) : OverdueEvent() {
+data class OverduePatientClicked(val patientUuid: UUID) : OverdueEvent() {
   override val analyticsName = "Overdue Screen:Patient name clicked"
 }