+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Changes
 - Updated translations: `pa-IN`, `hi-IN`, `te-IN`, `kn-IN`
+- Click on overdue patient to open patient summary  
 - [In Progress: 17 Jun 2021] Overdue list improvements
   - Change overdue list UI
   - Load overdue patients without phone number